# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/catalog": {
      "get": {
        "operationId": "AdminCatalogController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatalogCategoryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin · catalog"
        ]
      }
    },
    "/v1/admin/catalog/categories": {
      "post": {
        "operationId": "AdminCatalogController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogCategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin · catalog"
        ]
      }
    },
    "/v1/admin/catalog/categories/{id}": {
      "patch": {
        "operationId": "AdminCatalogController_updateCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogCategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin · catalog"
        ]
      },
      "delete": {
        "operationId": "AdminCatalogController_deleteCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin · catalog"
        ]
      }
    },
    "/v1/admin/catalog/items": {
      "post": {
        "operationId": "AdminCatalogController_createItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin · catalog"
        ]
      }
    },
    "/v1/admin/catalog/items/{id}": {
      "patch": {
        "operationId": "AdminCatalogController_updateItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin · catalog"
        ]
      },
      "delete": {
        "operationId": "AdminCatalogController_deleteItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletes the item, or deactivates it if past orders reference it.",
        "tags": [
          "admin · catalog"
        ]
      }
    },
...
    "/v1/orders/quote": {
      "post": {
        "operationId": "OrdersController_quote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Server-side price for a basket, including promo validation. Call before placing.",
        "tags": [
          "orders"
        ]
      }
    },
    "/v1/orders": {
      "post": {
        "operationId": "OrdersController_place",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "One value per checkout attempt (a UUID). Retrying with the same key returns the order the first attempt placed, with the Idempotent-Replayed: true header, instead of placing another.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "description": "active = not yet delivered or cancelled; past = delivered or cancelled; omit for all",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "past"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
...
    "/v1/orders/{id}/cancel": {
      "post": {
        "operationId": "OrdersController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Allowed until the clothes are picked up.",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/v1/admin/orders/{id}": {
      "get": {
        "operationId": "AdminOrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin · orders"
        ]
      }
    },
...
    "/v1/driver/orders/{id}": {
      "get": {
        "operationId": "DriverController_order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "driver"
        ]
      }
    },
...
  },
  "info": {
    "title": "Laundry API",
    "description": "Firebase phone-OTP ID token as Bearer. Money in paise. Dates are IST (YYYY-MM-DD).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ItemUnit": {
        "type": "string",
        "enum": [
          "PIECE",
          "PAIR",
          "SET",
          "KG"
        ]
      },
      "CatalogItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ItemUnit"
              }
            ]
          },
          "pricePaise": {
            "type": "number"
          },
          "offerPricePaise": {
            "type": "number",
            "nullable": true
          },
          "effectivePricePaise": {
            "type": "number",
            "description": "What the customer pays per unit."
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "categoryId",
          "name",
          "description",
          "unit",
          "pricePaise",
          "offerPricePaise",
          "effectivePricePaise",
          "imageUrl",
          "sortOrder",
          "isActive"
        ]
      },
      "CatalogCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogItemDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "imageUrl",
          "sortOrder",
          "isActive",
          "items"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ironing"
          },
          "slug": {
            "type": "string",
            "example": "ironing"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "default": 0
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ironing"
          },
          "slug": {
            "type": "string",
            "example": "ironing"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "default": 0
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "CreateItemDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Shirt"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "default": "PIECE",
            "allOf": [
              {
                "$ref": "#/components/schemas/ItemUnit"
              }
            ]
          },
          "pricePaise": {
            "type": "number",
            "example": 1500,
            "description": "Price in paise (₹15.00 = 1500)"
          },
          "offerPricePaise": {
            "type": "number",
            "nullable": true,
            "example": 1200,
            "description": "Discounted price in paise; must be below pricePaise"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "default": 0
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "categoryId",
          "name",
          "pricePaise"
        ]
      },
      "UpdateItemDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Shirt"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "default": "PIECE",
            "allOf": [
              {
                "$ref": "#/components/schemas/ItemUnit"
              }
            ]
          },
          "pricePaise": {
            "type": "number",
            "example": 1500,
            "description": "Price in paise (₹15.00 = 1500)"
          },
          "offerPricePaise": {
            "type": "number",
            "nullable": true,
            "example": 1200,
            "description": "Discounted price in paise; must be below pricePaise"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "default": 0
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "DeleteResultDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "description": "false when the record was kept but deactivated because orders use it"
          }
        },
        "required": [
          "deleted"
        ]
      },
...
      "QuoteRequestDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineInput"
            }
          },
          "promoCode": {
            "type": "string",
            "example": "FIRST50"
          }
        },
        "required": [
          "items"
        ]
      },
...
      "QuoteDto": {
        "type": "object",
        "properties": {
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteLineDto"
            }
          },
          "subtotalPaise": {
            "type": "number"
          },
          "discountPaise": {
            "type": "number"
          },
          "deliveryFeePaise": {
            "type": "number"
          },
          "totalPaise": {
            "type": "number"
          },
          "promoCode": {
            "type": "string",
            "nullable": true
          },
          "promoError": {
            "type": "string",
            "nullable": true,
            "description": "Why the promo code was not applied, if it was not.",
            "enum": [
              "PROMO_NOT_FOUND",
              "PROMO_EXPIRED",
              "PROMO_LIMIT_REACHED",
              "PROMO_MIN_ORDER"
            ]
          },
          "promoShortfallPaise": {
            "type": "number",
            "nullable": true
          },
          "minOrderShortfallPaise": {
            "type": "number",
            "nullable": true
          },
          "canPlaceOrder": {
            "type": "boolean"
          }
        },
        "required": [
          "lines",
          "subtotalPaise",
          "discountPaise",
          "deliveryFeePaise",
          "totalPaise",
          "promoCode",
          "promoError",
          "promoShortfallPaise",
          "minOrderShortfallPaise",
          "canPlaceOrder"
        ]
      },
...
      "PlaceOrderDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineInput"
            }
          },
          "promoCode": {
            "type": "string",
            "example": "FIRST50"
          },
          "pickupDate": {
            "type": "string",
            "example": "2026-10-05",
            "description": "IST calendar date"
          },
          "pickupSlot": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeSlot"
              }
            ]
          },
          "deliveryDate": {
            "type": "string",
            "example": "2026-10-06"
          },
          "deliverySlot": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeSlot"
              }
            ]
          },
          "pickupAddressId": {
            "type": "string"
          },
          "deliveryAddressId": {
            "type": "string",
            "description": "Defaults to the pickup address"
          },
          "paymentMethod": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentMethod"
              }
            ]
          },
          "instructions": {
            "type": "string",
            "example": "Please call before arriving"
          }
        },
        "required": [
          "items",
          "pickupDate",
          "pickupSlot",
          "deliveryDate",
          "deliverySlot",
          "pickupAddressId",
          "paymentMethod"
        ]
      },
...
      "OrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderNumber": {
            "type": "string",
            "example": "LD001042"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderStatus"
              }
            ]
          },
          "source": {
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderSource"
              }
            ]
          },
          "pickupDate": {
            "type": "string"
          },
          "pickupSlot": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeSlot"
              }
            ]
          },
          "pickupSlotLabel": {
            "type": "string"
          },
          "deliveryDate": {
            "type": "string"
          },
          "deliverySlot": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeSlot"
              }
            ]
          },
          "deliverySlotLabel": {
            "type": "string"
          },
          "pickupAddress": {
            "$ref": "#/components/schemas/OrderAddressDto"
          },
          "deliveryAddress": {
            "$ref": "#/components/schemas/OrderAddressDto"
          },
          "instructions": {
            "type": "string",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          },
          "subtotalPaise": {
            "type": "number"
          },
          "discountPaise": {
            "type": "number"
          },
          "deliveryFeePaise": {
            "type": "number"
          },
          "totalPaise": {
            "type": "number"
          },
          "paidPaise": {
            "type": "number"
          },
          "refundedPaise": {
            "type": "number",
            "description": "Sum of PENDING and PROCESSED refunds."
          },
          "refundablePaise": {
            "type": "number",
            "description": "What can still be refunded (paidPaise - refundedPaise). Admin view only."
          },
          "amountDuePaise": {
            "type": "number"
          },
          "promoCode": {
            "type": "string",
            "nullable": true
          },
          "paymentMethod": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentMethod"
              }
            ]
          },
          "paymentStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentStatus"
              }
            ]
          },
          "customer": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PersonRefDto"
              }
            ]
          },
          "pickupDriver": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PersonRefDto"
              }
            ]
          },
          "deliveryDriver": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PersonRefDto"
              }
            ]
          },
          "allowedNextStatuses": {
            "type": "array",
            "description": "Statuses the caller is allowed to move this order to.",
            "items": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          },
          "dispatchFailedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "cancelReason": {
            "type": "string",
            "nullable": true
          },
          "pickedUpAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "deliveredAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "cancelledAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderEventDto"
            }
          },
          "refunds": {
            "description": "Oldest first. Customer and admin views only.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RefundDto"
            }
          }
        },
        "required": [
          "id",
          "orderNumber",
          "status",
          "source",
          "pickupDate",
          "pickupSlot",
          "pickupSlotLabel",
          "deliveryDate",
          "deliverySlot",
          "deliverySlotLabel",
          "pickupAddress",
          "deliveryAddress",
          "instructions",
          "items",
          "subtotalPaise",
          "discountPaise",
          "deliveryFeePaise",
          "totalPaise",
          "paidPaise",
          "refundedPaise",
          "amountDuePaise",
          "promoCode",
          "paymentMethod",
          "paymentStatus",
          "customer",
          "pickupDriver",
          "deliveryDriver",
          "allowedNextStatuses",
          "dispatchFailedAt",
          "cancelReason",
          "pickedUpAt",
          "deliveredAt",
          "cancelledAt",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}